Rename scanUserEvents to scanUserEventsAndGeneratePlan
